This is the initial commit of a project that has been in development for a couple months. I should've been using git since the start, but I did not know my project's directory structure at the outset. It will be rearranged after this commit, such that everything is first saved, but as of now, the general structure:
1) The top-level has some domain-task-abstraction resources, i.e. resources to be able to
build something appropriate for the actual requirements of utility. The Archive Sheets
are the rough-input to the database, to-be-accomplished by an application-interface, likely
RStudio. The other half of this project, its complementary RShiny-app, is in another repo
located on this account.

2) Various notes, in .rtf's, as part of the domain task-abstraction mentioned above.

3) Within the mySQL folder:

a) top-level has some various scripts that are not in-use, but of some relevance for development.
b) "Model/" has had several iterations, and has been the bulk of the work thus far. There are 9
saved versions, but after this commit, all but the latest will be deleted, and versioning will
simply be tracked by GIT. So far the schema is in a mix of matching the current situation, and
exploring some changes that would perhaps be best practice. Some stub tables exist, such as
"person-audit" and "movie-audit", just to be visible during throughout the development process,
but will likely only be implemented at a later time.
c) "Scripts/" contain the frameworks, or beginnings of the core database facilities provided
through the RShiny-app interface. For security, integrity, maintenance, and convenience reasons,
the app will only be given execution-privilege under its login credentials.